+++ src/com/nvlad/yii2support/ErrorReportHandler.java
@@ -1,5 +1,6 @@
 package com.nvlad.yii2support;
 
+import com.intellij.diagnostic.AbstractMessage;
 import com.intellij.openapi.diagnostic.ErrorReportSubmitter;
 import com.intellij.openapi.diagnostic.IdeaLoggingEvent;
 import com.intellij.openapi.diagnostic.SubmittedReportInfo;
@@ -10,7 +11,7 @@ import org.jetbrains.annotations.Nullable;
 import java.awt.*;
 
 public class ErrorReportHandler extends ErrorReportSubmitter {
-
+    @NotNull
     @Override
     public String getReportActionText() {
         return "Report to Yii2 Support";
@@ -19,9 +20,13 @@ public class ErrorReportHandler extends ErrorReportSubmitter {
     @Override
     public boolean submit(@NotNull IdeaLoggingEvent[] events, @Nullable String additionalInfo, @NotNull Component parentComponent, @NotNull Consumer<SubmittedReportInfo> consumer) {
         for (IdeaLoggingEvent event : events) {
-            PluginApplicationComponent.getInstance().getRollbar().error(event.getThrowable(), additionalInfo);
-        }
+            Throwable throwable = event.getThrowable();
+            if (event.getData() instanceof AbstractMessage) {
+                throwable = ((AbstractMessage) event.getData()).getThrowable();
+            }
 
+            PluginApplicationComponent.getInstance().submitErrorReport(throwable, additionalInfo, consumer);
+        }
         return true;
     }
 }