+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Yii2 Support Change Log
 Waiting for Release
 -------------------
 - Show validators with equal class name
+- Fix submit error report for PhpStorm 2018.2
+- Enhance error report submitting
 
 0.10.56.15 August 13, 2018
 --------------------------